myArray + interface
