--- src/Main.java
@@ -1,18 +1,0 @@
-public class Main {
-    public static void main(String[] args) {
-      MyArrayList<Integer> test = new MyArrayList<>();
-      test.add(4);
-      test.add(6);
-      test.add(99);
-      test.add(33);
-        System.out.println(test.get(1));
-        test.set(1,66);
-        System.out.println(test.get(1));
-    }
-}
-
-/*
-2.1 Реализовать ArrayList Написать свой класс MyArrayList как аналог классу ArrayList.
- Можно использовать 1 массив для хранения данных.
-Реализовать следующие методы: add, remove, clear, size, get;
- */
