+++ src/com/homework/MyArrayList.java
@@ -1,0 +1,31 @@
+package com.homework;
+
+public class MyArrayList<E> implements MyList<E> {
+    private int size;
+
+
+    @Override
+    public void add(E e) {
+        
+    }
+
+    @Override
+    public void remove(int index) {
+
+    }
+
+    @Override
+    public void clear() {
+
+    }
+
+    @Override
+    public int size() {
+        return 0;
+    }
+
+    @Override
+    public E get(int index) {
+        return null;
+    }
+}
